작업완료
- PLL이 정상 동작하지 않는 이슈가 있어 추가 검토 필요

+++ yss/inc/drv/clock/drv_st_clock_type_F.h
@@ -173,11 +173,11 @@ class Clock
     bool enableLse(bool en = true);
     bool setUsbClkSrc(unsigned char src);
     bool setSysclk(unsigned char sysclkSrc, unsigned char ahb, unsigned char apb1, unsigned char apb2);
-    unsigned long getSysClkFreq(void);
-    unsigned long getApb1ClkFreq(void);
-    unsigned long getApb2ClkFreq(void);
-    unsigned long getTimerApb1ClkFreq(void);
-    unsigned long getTimerApb2ClkFreq(void);
+    unsigned int getSysClkFreq(void);
+    unsigned int getApb1ClkFreq(void);
+    unsigned int getApb2ClkFreq(void);
+    unsigned int getTimerApb1ClkFreq(void);
+    unsigned int getTimerApb2ClkFreq(void);
     Mainpll pll;
     Saipll saipll;
     Peripheral peripheral;
